requirements.txt
appdirs==1.4.3
astroid
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.3
backcall==0.1.0
certifi==2018.10.15
Click==7.0
decorator==4.3.0
gabrieltool==0.0.1
isort==4.3.4
lazy-object-proxy==1.3.1
logzero==1.5.0
mccabe==0.6.1
more-itertools==4.3.0
numpy==1.15.4
opencv-python==3.4.3.18
parso==0.3.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
py==1.7.0
pycodestyle==2.4.0
Pygments==2.2.0
pytest==4.0.0
rope==0.11.0
six==1.11.0
toml==0.10.0
traitlets==4.3.2
wcwidth==0.1.7
wrapt==1.10.11
